% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matched_set_obj.R
\name{balance_scatter}
\alias{balance_scatter}
\title{Visualizing the standardized mean differences for covariates}
\usage{
balance_scatter(non_refined_set, refined_list, xlim = c(0, 0.8), ylim = c(0,
  0.8), main = "Standardized Mean Difference of Covariates", pchs = c(2, 3),
  covariates, data, x.axis.label = "Before refinement",
  y.axis.label = "After refinement")
}
\arguments{
\item{non_refined_set}{a matched.set object produced by setting `refinement.method` to "none" in `PanelMatch`}

\item{refined_list}{a list of one or two matched.set objects}

\item{xlim}{xlim of the scatter plot. This is the same as the \code{xlim} argument in \code{plot}}

\item{ylim}{ylim of the scatter plot. This is the same as the \code{ylim} argument in \code{plot}}

\item{main}{title of the scatter plot. This is the same as the \code{main} argument in \code{plot}}

\item{pchs}{one or two pch indicators for the symbols on the scatter plot. See \code{plot} for more information}

\item{covariates}{variables for which balance is displayed}

\item{data}{the same time series cross sectional data set used to create the matched sets.}

\item{x.axis.label}{x axis label}

\item{y.axis.label}{y axis label}
}
\description{
\code{balance_scatter} visualizes the standardized mean differences for each covariate
}
\examples{
\dontrun{
sets1 <- PanelMatch(lag = 4, time.id = "year", unit.id = "wbcode2",
                    treatment = "dem", refinement.method = "CBPS.match",
                    data = dem, match.missing = F,
                    covs.formula = ~ I(lag(y, 1:4)) + I(lag(tradewb, 1:4)),
                    size.match = 5, qoi = "att",
                    outcome.var = "y",
                    lead = 0:4, forbid.treatment.reversal = FALSE)

sets0 <- PanelMatch(lag = 4, time.id = "year", unit.id = "wbcode2",
                    treatment = "dem", refinement.method = "none",
                    data = dem, match.missing = F,
                    covs.formula = ~ I(lag(y, 1:4)) + I(lag(tradewb, 1:4)),
                    size.match = 5, qoi = "att",
                    outcome.var = "y",
                    lead = 0:4, forbid.treatment.reversal = FALSE)

sets2 <- PanelMatch(lag = 4, time.id = "year", unit.id = "wbcode2",
                    treatment = "dem", refinement.method = "CBPS.weight",
                    data = dem, match.missing = F,
                    covs.formula = ~ I(lag(y, 1:4)) + I(lag(tradewb, 1:4)),
                    size.match = 10, qoi = "att",
                    outcome.var = "y",
                    lead = 0:4, forbid.treatment.reversal = FALSE)



balance_scatter(non_refined_set = sets0$att,
              refined_list = list(sets1$att, sets2$att),
              data = dem,
              covariates = c("y", "tradewb"))

}

}
\author{
In Song Kim <insong@mit.edu>, Erik Wang
<haixiao@Princeton.edu>, Adam Rauh <adamrauh@mit.edu>, and Kosuke Imai <imai@harvard.edu>
}
