% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matched_set_obj.R
\name{build_balance_mats}
\alias{build_balance_mats}
\title{#' @export
{
  
  # if(is.null(sets)) stop("please specify matched sets")
  # class(x) <- "data.frame"
  # ids <- x[i, c("i")]
  # ts <- x[i, c("t")]
  # 
  # attr(temp, "lag") <- attr(x, "lag")
  # attr(temp, "refinement.method") <- attr(x, "refinement.method")
  # class(temp) <- "matched.set"
  # return(temp)
}}
\usage{
build_balance_mats(idx, ordered_expanded_data, msets)
}
\description{
#' @export
{
  
  # if(is.null(sets)) stop("please specify matched sets")
  # class(x) <- "data.frame"
  # ids <- x[i, c("i")]
  # ts <- x[i, c("t")]
  # 
  # attr(temp, "lag") <- attr(x, "lag")
  # attr(temp, "refinement.method") <- attr(x, "refinement.method")
  # class(temp) <- "matched.set"
  # return(temp)
}
}
