% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_refinement.R
\name{build_maha_mats}
\alias{build_maha_mats}
\title{build_maha_mats
Builds the matrices that we will then use to calculate the mahalanobis distances for each matched set}
\usage{
build_maha_mats(idx, ordered_expanded_data)
}
\arguments{
\item{idx}{List of vectors specifying which observations should be extracted}

\item{ordered_expanded_data}{data.frame of prepared/parsed input data}
}
\value{
List of parsed distance matrices, with elements corresponding to each matched set
}
\description{
build_maha_mats
Builds the matrices that we will then use to calculate the mahalanobis distances for each matched set
}
\keyword{internal}
