% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity_score_utilities.R
\name{build_ps_data}
\alias{build_ps_data}
\title{build_ps_data}
\usage{
build_ps_data(idxlist, data, lag)
}
\arguments{
\item{idxlist}{}

\item{data}{data.frame object with the data}

\item{lag}{see PanelMatch() documentation}
}
\value{
Returns a list of length equal to the number of matched sets. Each item is a data frame and each data frame contains information at time = t + 0 for each treated unit and their corresponding controls.
}
\description{
build_ps_data
}
\keyword{internal}
