% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_placebo_estimates.R
\name{calculate_placebo_estimates}
\alias{calculate_placebo_estimates}
\title{calculate_placebo_estimates}
\usage{
calculate_placebo_estimates(
  qoi.in,
  data.in,
  lead,
  number.iterations,
  att.treated.unit.ids,
  atc.treated.unit.ids,
  outcome.variable,
  unit.id.variable,
  confidence.level,
  att.sets,
  atc.sets,
  placebo.test = FALSE,
  lag,
  placebo.lead,
  se.method = "bootstrap",
  parallel = FALSE,
  num.cores = 1
)
}
\value{
Returns a PanelEstimate object
}
\description{
Handles the procedures for calculating point estimates and standard errors for the placebo test. Code is structured very similarly to the calculate_estimates() code, but with appropriate modifications for the placebo test. See that function for description of arguments. Bootstrap SEs are available for any specification. Conditional, unconditional standard errors only available for att, art, atc.
}
\keyword{internal}
