% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_estimates.R
\name{calculate_point_estimates}
\alias{calculate_point_estimates}
\title{calculate_point_estimates
Helper function that calculates the point estimates for the specified QOI}
\usage{
calculate_point_estimates(
  qoi.in,
  data.in,
  lead,
  outcome.variable,
  pooled = FALSE
)
}
\arguments{
\item{qoi.in}{string specifying the QOI}

\item{data.in}{data.frame providing the processed/parsed data to be used for calculations}

\item{lead}{see PanelMatch() documentation}

\item{outcome.variable}{string specifying the outcome variable}

\item{pooled}{Logical. See PanelEstimate() documentation.}
}
\value{
A named vector of point estimates
}
\description{
calculate_point_estimates
Helper function that calculates the point estimates for the specified QOI
}
\keyword{internal}
