% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{check_time_data}
\alias{check_time_data}
\title{check_time_data}
\usage{
check_time_data(data, time.id)
}
\arguments{
\item{data}{data.frame object.}

\item{time.id}{string specifying the time id variable.}
}
\value{
data.frame object with the data. If function throws error, nothing is returned.
}
\description{
Time data should be consecutive integers: When it is not, try to convert it as best we can or throw an error. If function does not fail, returns the data as data frame object, either processed or not as appropriately
}
\details{
enforces the requirements for time data, with some reasonable defaults
}
\keyword{internal}
