% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_leads.R
\name{clean_leads}
\alias{clean_leads}
\title{clean_leads
Function to check the lead windows in treated and control units for missing outcome data. If data is missing, remove those units from matched sets.}
\usage{
clean_leads(matched_sets, ordered.data, max.lead, t.var, id.var, outcome.var)
}
\arguments{
\item{matched_sets}{matched.set object contained pre-filtered matched sets}

\item{ordered.data}{data.frame object to be checked for missing data. This should have been passed through data preparation functions already.}

\item{max.lead}{Integer specifying the biggest value of the lead window.}

\item{t.var}{string specifying the time id variable}

\item{id.var}{string specifying the unit id variable}

\item{outcome.var}{string specifying the outcome variable.}
}
\value{
a cleaned/filtered matched.set object
}
\description{
clean_leads
Function to check the lead windows in treated and control units for missing outcome data. If data is missing, remove those units from matched sets.
}
\keyword{internal}
