% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelEstimateObject.R
\name{confint.PanelEstimate}
\alias{confint.PanelEstimate}
\title{Produce confidence intervals for PanelEstimate objects}
\usage{
\method{confint}{PanelEstimate}(object, parm = NULL, level = NULL, ..., bias.corrected = FALSE)
}
\arguments{
\item{object}{\code{PanelEstimate} results}

\item{parm}{Not used.}

\item{level}{Confidence level to be used for confidence interval calculations. Must be numeric between 0 and 1. If NULL, confidence level from \code{PanelEstimate()} specification is used.}

\item{...}{not used}

\item{bias.corrected}{logical indicating whether or not bias corrected estimates should be provided. Default is FALSE. This argument only applies for standard errors calculated with the bootstrap.}
}
\value{
Matrix with two columns and `length(lead)` rows. Contains the upper and lower boundaries of the confidence interval for each time period's point estimate.
}
\description{
Produce confidence intervals for PanelEstimate objects
}
