% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matched_set_obj.R
\name{DifferenceScatter}
\alias{DifferenceScatter}
\title{DifferenceScatter}
\usage{
DifferenceScatter(non_refined_set, refined_list, xlim = c(0, 0.8),
  ylim = c(0, 0.8), main = "Standardized Mean Difference of Covariates",
  pchs = c(2, 3), covariates, data, x.axis.label = "Before refinement",
  y.axis.label = "After refinement")
}
\arguments{
\item{non_refined_set}{a matched.set object produced by setting `refinement.method` to "none" in `PanelMatch`}

\item{refined_list}{a list of one or two matched.set objects}

\item{xlim}{xlim of the scatter plot}

\item{ylim}{ylim of the scatter plot}

\item{main}{title of the scatter plot}

\item{pchs}{one or two pch for the symbols on the scatter plot}

\item{covariates}{variables for which balance is displayed}

\item{data}{the dataset}

\item{x.axis.label}{x axis label}

\item{y.axis.label}{y axis label}
}
\description{
\code{DifferenceScatter} visualizes the standardized mean differences for each covariate
}
\examples{
\dontrun{
sets1 <- PanelMatch(lag = 4, time.id = "year", unit.id = "wbcode2",
                    treatment = "dem", refinement.method = "CBPS.match",
                    data = dem, match.missing = F,
                    covs.formula = ~ lag("y", 1:4) + lag("tradewb", 1:4),
                    size.match = 5, qoi = "att",
                    outcome.var = "y",
                    lead = 4, forbid.treatment.reversal = FALSE)

sets0 <- PanelMatch(lag = 4, time.id = "year", unit.id = "wbcode2",
                    treatment = "dem", refinement.method = "none",
                    data = dem, match.missing = F,
                    covs.formula = ~ lag("y", 1:4) + lag("tradewb", 1:4),
                    size.match = 5, qoi = "att",
                    outcome.var = "y",
                    lead = 4, forbid.treatment.reversal = FALSE)

sets2 <- PanelMatch(lag = 4, time.id = "year", unit.id = "wbcode2",
                    treatment = "dem", refinement.method = "CBPS.weight",
                    data = dem, match.missing = F,
                    covs.formula = ~ lag("y", 1:4) + lag("tradewb", 1:4),
                    size.match = 10, qoi = "att",
                    outcome.var = "y",
                    lead = 4, forbid.treatment.reversal = FALSE)



DifferenceScatter(non_refined_set = sets0$att,
              refined_lists = list(sets1$att, sets2$att),
              
              data = dem,
              covariates = c("y", "tradewb"))
}

}
\author{
In Song Kim <insong@mit.edu>, Erik Wang
<haixiao@Princeton.edu>, Adam Rauh <adamrauh@mit.edu>, and Kosuke Imai <kimai@Princeton.edu>
}
