% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisplayTreatment.R
\name{DisplayTreatment}
\alias{DisplayTreatment}
\title{DisplayTreatment}
\usage{
DisplayTreatment(unit.id, time.id, treatment, data,
color.of.treated, color.of.untreated, title, legend.position, xlab,
ylab, x.size, y.size, legend.labels)
}
\arguments{
\item{unit.id}{A numeric vector of unit identifiers}

\item{time.id}{A numeric vector of time identifiers}

\item{treatment}{Name of the treatment variable in character class}

\item{data}{The data frame in data.frame class}

\item{color.of.treated}{Color of the treated observations in
character. Default is red.}

\item{color.of.untreated}{Color of the untreated observations in
character. Defalt is blue.}

\item{title}{Title of the plot in character}

\item{legend.position}{Position of the legend with the same choice
set as in ggplot2}

\item{xlab}{Character label of the x-axis}

\item{ylab}{Character label of the y-axis}

\item{x.size}{Numeric size of the text for xlab. Default is 10}

\item{y.size}{Numeric size of the text for ylab. Default is 5}

\item{legend.labels}{Character vector of length two describing the
labels of the legend to be shown in the plot}
}
\value{
\code{DisplayTreatment} returns a treatment variation plot,
which visualizes the variation of treatment across unit and time.
}
\description{
\code{DisplayTreatment} visualizes the treatment distribution
across unit and time in a panel dataset
}
\examples{
\dontrun{
library(ggplot2)
data <- dem
DisplayTreatment(unit.id = "wbcode2",
                 time.id = "year", legend.position = "none",
                 xlab = "year", ylab = "Country Code",
                 treatment = "dem", data = data)
}
}
\author{
In Song Kim <insong@mit.edu>, Erik Wang
<haixiao@Princeton.edu>, and Kosuke Imai <kimai@Princeton.edu>
}
