% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matched_set_obj.R
\name{distances}
\alias{distances}
\title{Get distances
See distances.matched.set method}
\usage{
distances(object)
}
\arguments{
\item{object}{\code{matched.set} object}
}
\description{
Get distances
See distances.matched.set method
}
