% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enforce_lead_restrictions.R
\name{enforce_lead_restrictions}
\alias{enforce_lead_restrictions}
\title{enforce_lead_restrictions
check treatment and control units for treatment reversion in the lead window. Treated units must stay treated and control units must stay in control (according to the specified qoi)}
\usage{
enforce_lead_restrictions(
  matched_sets,
  ordered.data,
  max.lead,
  t.var,
  id.var,
  treatment.var
)
}
\arguments{
\item{matched_sets}{matched.set object}

\item{ordered.data}{parsed data as data.frame object}

\item{max.lead}{The largest lead value (e.g. the biggest F)}

\item{t.var}{string specifying the time variable}

\item{id.var}{string specifying the unit id variable}

\item{treatment.var}{string specifying the treatment variable.}
}
\value{
matched.set object with the matched sets that meet the conditions
}
\description{
enforce_lead_restrictions
check treatment and control units for treatment reversion in the lead window. Treated units must stay treated and control units must stay in control (according to the specified qoi)
}
