% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_estimates.R
\name{equality_four}
\alias{equality_four}
\title{equality_four
Small helper function implementing estimation function from Imai, Kim, and Wang (2023)}
\usage{
equality_four(x, y, z)
}
\value{
Returns numeric vector of results.
}
\description{
equality_four
Small helper function implementing estimation function from Imai, Kim, and Wang (2023)
}
\keyword{internal}
