% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_placebo_estimates.R
\name{equality_four_placebo}
\alias{equality_four_placebo}
\title{equality_four_placebo}
\usage{
equality_four_placebo(x, y, z)
}
\value{
Returns numeric vector of results.
}
\description{
Small helper function implementing estimation function from Imai, Kim, and Wang (2023)
}
\keyword{internal}
