% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelEstimateObject.R
\name{estimates.PanelEstimate}
\alias{estimates.PanelEstimate}
\title{Extract QOI estimates}
\usage{
\method{estimates}{PanelEstimate}(object, ...)
}
\arguments{
\item{object}{\code{PanelEstimate} object}

\item{...}{not used}
}
\value{
Named vector with the QOI point estimates and the time periods to which they correspond
}
\description{
This is a method for extracting point estimates for the QOI from \code{PanelEstimate} objects. This function is analogous to the `coef()` method used elsewhere.
}
