% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelEstimateObject.R
\name{estimates}
\alias{estimates}
\title{Extract QOI estimates
See documentation for `estimates.PanelEstimate()`}
\usage{
estimates(object, ...)
}
\arguments{
\item{object}{}

\item{...}{}
}
\description{
Extract QOI estimates
See documentation for `estimates.PanelEstimate()`
}
