% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matched_set_helpers.r
\name{expand_treated_ts}
\alias{expand_treated_ts}
\title{expand_treated_ts
Builds a list that contains all times in a lag window that correspond to a particular treated unit. This is structured as a list of vectors. Each vector is lag + 1 units long. The overall list will be the same length as the number of matched sets}
\usage{
expand_treated_ts(lag, treated.ts)
}
\arguments{
\item{lag}{lag value}

\item{treated.ts}{times of treated observations}
}
\value{
list. Contains all times in a lag window that correspond to a particular treated unit
}
\description{
expand_treated_ts
Builds a list that contains all times in a lag window that correspond to a particular treated unit. This is structured as a list of vectors. Each vector is lag + 1 units long. The overall list will be the same length as the number of matched sets
}
\keyword{internal}
