% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelMatchObject.R
\name{extract}
\alias{extract}
\title{Extract matched.set objects from PanelMatch results}
\usage{
extract(pm.object, qoi)
}
\arguments{
\item{pm.object}{\code{PanelMatch} object}

\item{qoi}{character, specifying the qoi. Valid inputs include "att", "atc", "art", and NUL}
}
\description{
Extract matched.set objects from PanelMatch results
}
