% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matched_set_helpers.r
\name{extract_differences}
\alias{extract_differences}
\title{extract_differences
This function calculates the differences from t-1 to 1 for treated and control units in the treatment variable. While functionality is somewhat trivial for current implementation of package, it will be needed for continuous treatment version of the package.}
\usage{
extract_differences(indexed.data, matched.set, treatment.variable, qoi)
}
\arguments{
\item{indexed.data}{data that has been indexed. Rows have been named with a unique identifier.}

\item{matched.set}{matched.set object}

\item{treatment.variable}{string specifying treatment variable}

\item{qoi}{string specifying QOI}
}
\value{
matched.set object, with differences extracted as described previously for each matched set.
}
\description{
extract_differences
This function calculates the differences from t-1 to 1 for treated and control units in the treatment variable. While functionality is somewhat trivial for current implementation of package, it will be needed for continuous treatment version of the package.
}
\keyword{internal}
