% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelMatchObject.R
\name{formula.PanelMatch}
\alias{formula.PanelMatch}
\title{Return the refinement formula used in a PanelMatch specification}
\usage{
\method{formula}{PanelMatch}(x, ...)
}
\arguments{
\item{...}{not used}

\item{A}{PanelMatch Object}
}
\value{
One sided formula object containing the variables/specification used in refinement. This corresponds to what was provided to the \code{covs.formula} argument.
}
\description{
Return the refinement formula used in a PanelMatch specification
}
