% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelBalance.R
\name{get_unrefined_balance.PanelBalance}
\alias{get_unrefined_balance.PanelBalance}
\title{Extract unrefined covariate balance results, if they exist}
\usage{
\method{get_unrefined_balance}{PanelBalance}(pb.object)
}
\arguments{
\item{pb.object}{PanelBalance object}
}
\value{
A PanelBalance object, with just the unrefined balance results
}
\description{
Extract unrefined covariate balance results, if they exist
}
