% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelBalance.R
\name{get_unrefined_balance}
\alias{get_unrefined_balance}
\title{Extract just the unrefined covariate balance results, if they exist}
\usage{
get_unrefined_balance(pb.object)
}
\arguments{
\item{pb.object}{\code{PanelBalance} object}
}
\description{
Extract just the unrefined covariate balance results, if they exist
}
