% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matched_set_obj.R
\name{getSetTreatmentEffects}
\alias{getSetTreatmentEffects}
\title{get_set_treatment_effects}
\usage{
getSetTreatmentEffects(pm.obj, data.in, lead)
}
\arguments{
\item{pm.obj}{an object of class \code{PanelMatch}}

\item{data.in}{data.frame with the original data}

\item{lead}{integer (or integer vector) indicating the time period(s) in the future for which the treatment effect size will be calculated. Calculations will be made for the period t + lead, where t is the time of treatment. If more than one lead value is provided, then calculations will be performed for each value.}
}
\description{
Calculates the treatment effect size at the matched set level
}
\details{
Calculate the size of treatment effects for each matched set.
}
\examples{
PM.results <- PanelMatch(lag = 4, time.id = "year", unit.id = "wbcode2",
set.effects <- getSetTreatmentEffects(pm.obj = PM.results, data.in = dem, lead = 0)
}
