% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_components.R
\name{getWits}
\alias{getWits}
\title{getWits
returns a vector of Wits, as defined in the paper (equation 25 or equation 23). They should be in the same order as the data frame containing the original problem data. The pts, pcs, and getWits functions act for a specific lead. So, for instance if our lead window is 0,1,2,3,4, these function must be called for each of those -- so for 0, then for 1, etc.}
\usage{
getWits(matched_sets, lead, data, estimation.method = "bootstrap")
}
\arguments{
\item{matched_sets}{matched.set object}

\item{lead}{integer providing a specific lead value}

\item{data}{data.frame object}

\item{estimation.method}{method of estimation for calculating standard errors.}
}
\value{
data.table of Wits, as described above
}
\description{
getWits
returns a vector of Wits, as defined in the paper (equation 25 or equation 23). They should be in the same order as the data frame containing the original problem data. The pts, pcs, and getWits functions act for a specific lead. So, for instance if our lead window is 0,1,2,3,4, these function must be called for each of those -- so for 0, then for 1, etc.
}
\keyword{internal}
