% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_errors.R
\name{handle_bootstrap}
\alias{handle_bootstrap}
\title{handle_bootstrap}
\usage{
handle_bootstrap(
  qoi.in,
  data.in,
  lead,
  number.iterations,
  att.treated.unit.ids,
  atc.treated.unit.ids,
  outcome.variable,
  unit.id.variable,
  confidence.level,
  lag,
  pooled
)
}
\arguments{
\item{qoi.in}{String specifying qoi}

\item{data.in}{data.frame object with the data}

\item{number.iterations}{integer. Specifies number of bootstrap iterations}

\item{att.treated.unit.ids}{Integer vector specifying the treated units for the att or art}

\item{atc.treated.unit.ids}{Integer vector specifying the "treated" units under the atc definition}

\item{outcome.variable}{string specifying the name of the outcome variable}

\item{unit.id.variable}{string specifying the name of the unit id variable}

\item{confidence.level}{double. specifies confidence level for confidence interval}

\item{lag}{integer vector specifying size of the lag.}

\item{pooled}{logical. Specifies whether or not to calculate point estimates for each specified lead value, or a single pooled estimate.}
}
\value{
Returns a matrix of bootstrapped QOI estimate values.
}
\description{
Helper function for calculating bootstrapped estimates for the QOI. This version is not parallelized.
}
\keyword{internal}
