% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_placebo.R
\name{handle_bootstrap_placebo}
\alias{handle_bootstrap_placebo}
\title{handle_bootstrap_placebo}
\usage{
handle_bootstrap_placebo(
  qoi.in,
  data.in,
  placebo.lead,
  number.iterations,
  att.treated.unit.ids,
  atc.treated.unit.ids,
  outcome.variable,
  unit.id.variable,
  confidence.level,
  lag
)
}
\arguments{
\item{qoi.in}{String specifying qoi}

\item{data.in}{data.frame object with the data}

\item{number.iterations}{integer. specifies number of bootstrap iterations}

\item{att.treated.unit.ids}{Integer vector specifying the treated units for the att or art}

\item{atc.treated.unit.ids}{Integer vector specifying the "treated" units under the atc definition}

\item{outcome.variable}{string specifying the name of the outcome variable}

\item{unit.id.variable}{string specifying the name of the unit id variable}

\item{confidence.level}{double. specifies confidence level for confidence interval}

\item{lag}{integer vector specifying size of the lag.}
}
\value{
Returns a matrix of bootstrapped QOI estimate values.
}
\description{
Helper function for calculating bootstrapped estimates for the placebo test. This version is not parallelized.
}
\keyword{internal}
