% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_refinement.R
\name{handle_mahalanobis_calculations}
\alias{handle_mahalanobis_calculations}
\title{handle_mahalanobis_calculations
Returns a matched.set object with weights for control units, along with some other metadata}
\usage{
handle_mahalanobis_calculations(
  mahal.nested.list,
  msets,
  max.size,
  verbose,
  use.diagonal.covmat
)
}
\arguments{
\item{mahal.nested.list}{Output from build_maha_mats function}

\item{msets}{matched.set object -- list containing the treated observations and matched controls}

\item{max.size}{maximum number of control units that will receive non-zero weights within a matched set}

\item{verbose}{Logical. See PanelMatch() documentation}

\item{use.diagonal.covmat}{Logical. See PanelMatch() documentation}
}
\value{
matched.set object with weights for control units, along with some other metadata
}
\description{
handle_mahalanobis_calculations
Returns a matched.set object with weights for control units, along with some other metadata
}
\keyword{internal}
