% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_moderating_variable.R
\name{handle_moderating_variable}
\alias{handle_moderating_variable}
\title{handle_moderating_variable
handles moderating variable calculations: In practice, this just involves slicing the data up according to the moderator, calling PanelEstimate() and putting everything back together This function creates the sets of objects on which PanelEstimate() will be called. It identifies the set of valid values the moderating variable can take on.}
\usage{
handle_moderating_variable(
  ordered.data,
  att.sets,
  atc.sets,
  PM.object,
  moderator,
  unit.id,
  time.id,
  qoi.in
)
}
\arguments{
\item{ordered.data}{data.frame}

\item{att.sets}{matched.set object for the ATT or ART}

\item{atc.sets}{matched.set object for the ATC}

\item{PM.object}{PanelMatch object}

\item{moderator}{string specifying the name of the moderating variable}

\item{unit.id}{string specifying the unit id variable}

\item{time.id}{string specifying the time id variable}

\item{qoi.in}{string specifying the QOI}
}
\value{
Character vector of valid moderating variable values
}
\description{
handle_moderating_variable
handles moderating variable calculations: In practice, this just involves slicing the data up according to the moderator, calling PanelEstimate() and putting everything back together This function creates the sets of objects on which PanelEstimate() will be called. It identifies the set of valid values the moderating variable can take on.
}
