% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_refinement.R
\name{handle_ps_match}
\alias{handle_ps_match}
\title{handle_ps_match
Returns a matched.set object with weights for control units, along with some other metadata}
\usage{
handle_ps_match(just.ps.sets, msets, refinement.method, verbose, max.set.size)
}
\arguments{
\item{just.ps.sets}{Output from find_ps() function}

\item{msets}{matched.set object -- list containing the treated observations and matched controls}

\item{verbose}{Logical. See PanelMatch() documentation}

\item{max.set.size}{maximum number of control units that will receive non-zero weights within a matched set}
}
\value{
matched.set object with weights for control units, along with some other metadata
}
\description{
handle_ps_match
Returns a matched.set object with weights for control units, along with some other metadata
}
\keyword{internal}
