% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting_refinement.R
\name{handle_ps_weighted}
\alias{handle_ps_weighted}
\title{handle_ps_weighted}
\usage{
handle_ps_weighted(just.ps.sets, msets, refinement.method)
}
\arguments{
\item{just.ps.sets}{results of find_ps()}

\item{msets}{list of matched sets of treated and control observations}

\item{refinement.method}{string specifying the refinement method}
}
\value{
matched.set object with treated and matched control observations, with weights as determined by the specification
}
\description{
handle_ps_weighted
}
\keyword{internal}
