% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_errors.R
\name{handle_unconditional_se}
\alias{handle_unconditional_se}
\title{handle_conditional_se
Calculates conditional standard errors analytically, as defined in Imai et al. (2023). See PanelEstimate() for a more complete description of the standard error types.}
\usage{
handle_unconditional_se(
  qoi.in,
  data.in,
  lead,
  outcome.variable,
  unit.id.variable
)
}
\arguments{
\item{qoi.in}{string specifying the QOI}

\item{data.in}{data.frame specifying the data}

\item{lead}{See PanelMatch() documentation}

\item{outcome.variable}{string specifying the name of the outcome variable}

\item{unit.id.variable}{string specifying the name of the unit id variable.}
}
\value{
Named vector with standard error estimates
}
\description{
handle_conditional_se
Calculates conditional standard errors analytically, as defined in Imai et al. (2023). See PanelEstimate() for a more complete description of the standard error types.
}
\keyword{internal}
