% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matched_set_helpers.r
\name{identifyDirectionalChanges}
\alias{identifyDirectionalChanges}
\title{identifyDirectionalChanges
Identifies changes in treatment variable for treated and control observations}
\usage{
identifyDirectionalChanges(
  msets,
  ordered.data,
  id.var,
  time.var,
  treatment.var,
  qoi
)
}
\arguments{
\item{msets}{}

\item{ordered.data}{}

\item{id.var}{}

\item{time.var}{}

\item{treatment.var}{}

\item{qoi}{}
}
\value{
matched.set object with changes in the treatment variable for treated and control observations identified.
}
\description{
identifyDirectionalChanges
Identifies changes in treatment variable for treated and control observations
}
\keyword{internal}
