% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listwise_delete.R
\name{lwd_units}
\alias{lwd_units}
\title{lwd_units
helper function that actually subsets sets down to contain units with complete data}
\usage{
lwd_units(full.local.data, unit.id)
}
\arguments{
\item{full.local.data}{data.frame containing the data to be used in set-level refinement, but containing missing data}

\item{unit.id}{}
}
\value{
data.frame with the missing data removed to be used for set-level refinement.
}
\description{
lwd_units
helper function that actually subsets sets down to contain units with complete data
}
\keyword{internal}
