% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_covs_formula.R
\name{merge_formula}
\alias{merge_formula}
\title{merge_formula}
\usage{
merge_formula(form1, form2)
}
\arguments{
\item{form1}{formula object}

\item{form2}{formula object}
}
\value{
Returns a formula object, which is the concatenation of two provided formula objects.
}
\description{
Simple helper function for merging formula objects
}
\keyword{internal}
