% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelEffectPlot.R
\name{PanelEffectPlot}
\alias{PanelEffectPlot}
\title{PanelEffectPlot}
\usage{
PanelEffectPlot(result, CI = 0.95, atc.reverse = FALSE,
  bias.correction = TRUE, xlab = "Effect by Time Period",
  ylab = "Effect Size", x.size = 13, y.size = 15, title = "",
  point_size = 2)
}
\arguments{
\item{result}{An object returned by PanelEstimate}

\item{CI}{A numeric value indicating the confidence interval set by the user. 
By default, it is .95}

\item{atc.reverse}{A logical value indicating whether the user wants to reverse the signs 
for effects when the quantity of interest is "atc" in PanelEstimate}

\item{bias.correction}{A logical value indicating whether bias correction standard errors
are used.}

\item{xlab}{A character string for x-axis label}

\item{ylab}{A character string for y-axis label}

\item{x.size}{A numeric value indicating the font size for the x-axis label}

\item{y.size}{A numeric value indicating the font size for the y-axis label}

\item{title}{A character string for the title of the plot}

\item{point_size}{A numeric value indicating the area of the circle that is 
going to appear in the plot to represent the point estimate for a given time
period.}
}
\value{
It returns a plot
}
\description{
\code{PanelEffectPlot} visualizes the results returned by \code{PanelEstimate} 
when the option \code{inference} in \code{PanelEstimate} is set to 
"bootstrap" via 
plotting point estimates along with their confidence intervals by time periods
specified by the user.
}
\examples{
\dontrun{
matches.cbps <- PanelMatch(lag = 4, max.lead = 4, time.id = "year",
unit.id = "wbcode2", treatment = "dem", formula = y ~ dem, method =
"CBPS", weighting = FALSE, qoi = "ate", M = 5, data = dem)

## bootstrap

mod.bootSE <- PanelEstimate(lead = 0:4, inference =
"bootstrap", matched_sets = matches.cbps, qoi = "att", CI = .95,
ITER = 500) 

PanelEffectPlot(mod.bootSE)

}
}
