% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelEstimate.R
\name{PanelEstimate}
\alias{PanelEstimate}
\title{Illustration of PanelEstimate
\code{PanelEstimate} estimates results by aggregating across adjusted matched sets obtained from \code{PanelMatch} via  weighted fixed effects regressions or via weighted average computation with weighted bootstrap standard errors}
\usage{
PanelEstimate(lead, covariate.only = FALSE, inference = c("wfe",
  "bootstrap"), ITER = 1000, matched_sets = NULL, plot = FALSE,
  estimator = "did", df.adjustment = FALSE, qoi = NULL, CI = 0.9)
}
\description{
Illustration of PanelEstimate
\code{PanelEstimate} estimates results by aggregating across adjusted matched sets obtained from \code{PanelMatch} via  weighted fixed effects regressions or via weighted average computation with weighted bootstrap standard errors
}
