% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelEstimate_tmp2.R
\name{PanelEstimate_tmp2}
\alias{PanelEstimate_tmp2}
\title{Illustration of PanelEstimate_tmp2}
\usage{
PanelEstimate_tmp2(lead, covariate.only = FALSE, inference = c("wfe",
  "bootstrap"), ITER = 1000, matched_sets = NULL, plot = FALSE,
  estimator = "did", df.adjustment = FALSE, qoi = NULL, CI = 0.9)
}
\description{
Illustration of PanelEstimate_tmp2
}
