% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelMatch2.R
\name{PanelMatch2.matched.set}
\alias{PanelMatch2.matched.set}
\title{This is currently only used inside PE2, where it is used to update the weights of a matched set object after control units that are missing data in the necessary future or past periods have been identified and taken out.
Function assumes that the units that are included are the only units it needs to consider. Pulls argument information from matched.set object 
Major difference between this function and the actual PanelMatch2 function is that this one does not identify matched sets and/or treated units. It merely updates and refines the matched set object provided to it}
\usage{
PanelMatch2.matched.set(mset.object, data, outcome.var)
}
\description{
This is currently only used inside PE2, where it is used to update the weights of a matched set object after control units that are missing data in the necessary future or past periods have been identified and taken out.
Function assumes that the units that are included are the only units it needs to consider. Pulls argument information from matched.set object 
Major difference between this function and the actual PanelMatch2 function is that this one does not identify matched sets and/or treated units. It merely updates and refines the matched set object provided to it
}
