% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_covs_formula.R
\name{parse_and_prep}
\alias{parse_and_prep}
\title{parse_and_prep}
\usage{
parse_and_prep(formula, data)
}
\arguments{
\item{formula}{formula object specifying how to construct the data used for refinement. This is likely to be some variation of the covs.formula argument.}

\item{data}{data.frame object to be used to create the data needed for refinement. data has unit, time, treatment columns in that order, followed by everything else}
}
\value{
data.frame object with the data prepared for refinement. Data will have unit, time, treatment columns in that order, followed by everything else.
}
\description{
accepts formula object and data, creates the data used for refinement
}
\keyword{internal}
