% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_refinement.R
\name{perform_refinement}
\alias{perform_refinement}
\title{perform_refinement
Performs refinement of matched sets, ultimately returning sets of treated observations and controls with weights. This function mostly acts as an intermediary between PanelMatch and lower level functions that do the dirty work of refinement. The function takes a lot of the same arguments as PanelMatch()}
\usage{
perform_refinement(
  lag,
  time.id,
  unit.id,
  treatment,
  refinement.method,
  size.match,
  ordered.data,
  match.missing,
  covs.formula,
  verbose,
  lead,
  outcome.var = NULL,
  forbid.treatment.reversal = FALSE,
  qoi = "",
  matching = TRUE,
  exact.matching.variables = NULL,
  listwise.deletion,
  use.diag.covmat = FALSE,
  placebo.test = FALSE,
  restrict.control.period = NULL
)
}
\arguments{
\item{lag}{See PanelMatch() documentation.}

\item{time.id}{See PanelMatch() documentation.}

\item{unit.id}{See PanelMatch() documentation.}

\item{treatment}{See PanelMatch() documentation.}

\item{refinement.method}{See PanelMatch() documentation.}

\item{size.match}{See PanelMatch() documentation.}

\item{ordered.data}{data.frame that has been balanced and ordered by time-unit.}

\item{match.missing}{See PanelMatch() documentation.}

\item{covs.formula}{See PanelMatch() documentation.}

\item{verbose}{See PanelMatch() documentation.}

\item{lead}{See PanelMatch() documentation.}

\item{outcome.var}{See PanelMatch() documentation.}

\item{forbid.treatment.reversal}{See PanelMatch() documentation.}

\item{qoi}{See PanelMatch() documentation.}

\item{matching}{See PanelMatch() documentation.}

\item{exact.matching.variables}{}

\item{listwise.deletion}{See PanelMatch() documentation.}

\item{use.diag.covmat}{See PanelMatch() documentation for use.diagonal.covariance.matrix argument.}

\item{placebo.test}{See PanelMatch() documentation.}

\item{restrict.control.period}{See PanelMatch() documentation.}
}
\value{
returns a matched.set object containing the refined matched sets
}
\description{
perform_refinement
Performs refinement of matched sets, ultimately returning sets of treated observations and controls with weights. This function mostly acts as an intermediary between PanelMatch and lower level functions that do the dirty work of refinement. The function takes a lot of the same arguments as PanelMatch()
}
\keyword{internal}
