% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_errors.R
\name{perunitSum}
\alias{perunitSum}
\title{perunitSum
This is a low level function that is used to calculate a value associated with each unit. This value is a weighted summation of the dependent variable, based on the Wit values discussed in Imai et al. (2023)}
\usage{
perunitSum(udf, lead.in, dependent.in, qoi_in)
}
\arguments{
\item{udf}{data.frame}

\item{lead.in}{integer. A particular lead value}

\item{dependent.in}{string specifying the dependent variable name}

\item{qoi_in}{string specifying the QOI}
}
\value{
Named vector containing the per-unit sums.
}
\description{
perunitSum
This is a low level function that is used to calculate a value associated with each unit. This value is a weighted summation of the dependent variable, based on the Wit values discussed in Imai et al. (2023)
}
