% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_errors.R
\name{perunitSum_Dit}
\alias{perunitSum_Dit}
\title{perunitSum_Dit
Similar to perunitSum, this is a low level helper function for calculating specific values defined in Imai et al. (2023). This focuses on Dit rather than Wit}
\usage{
perunitSum_Dit(udf, qoi_in)
}
\arguments{
\item{udf}{data.frame}

\item{qoi_in}{string specifying the QOI}
}
\value{
Named vector containing the per-unit sums.
}
\description{
perunitSum_Dit
Similar to perunitSum, this is a low level helper function for calculating specific values defined in Imai et al. (2023). This focuses on Dit rather than Wit
}
\keyword{internal}
