% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{placeboTest}
\alias{placeboTest}
\title{placeboTest}
\usage{
placeboTest(
  pm.obj,
  data.in,
  lag.in = NULL,
  number.iterations = 5000,
  df.adjustment = FALSE,
  confidence.level = 0.95,
  plot = FALSE,
  ...
)
}
\arguments{
\item{pm.obj}{an object of class \code{PanelMatch}}

\item{data.in}{data.frame with the original data}

\item{lag.in}{integer vector indicating the time period(s) in the future for which the placebo test change in outcome will be calculated. Calculations will be made over the period t - max(lag) to t-2, where t is the time of treatment. The results are similar to those returned by PanelEstimate, except t-1 is used as the period of comparison, rather than the lead window.}

\item{number.iterations}{integer specifying the number of bootstrap iterations}

\item{confidence.level}{confidence level for the calculated standard error intervals}

\item{plot}{logical indicating whether or not a plot should be generated, or just return the raw data from the calculations}
}
\description{
Calculates results for a placebo test
}
\details{
Calculate the results of a placebo test, looking at the change in outcome at time = t-1, compared to other pre treatment periods in the lag window.
}
