% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelBalance.R
\name{plot.PanelBalance}
\alias{plot.PanelBalance}
\title{plot covariate balance results}
\usage{
\method{plot}{PanelBalance}(
  x,
  ...,
  type = "panel",
  reference.line = TRUE,
  legend = TRUE,
  ylab = "SD",
  include.treatment.period = TRUE,
  include.unrefined.panel = TRUE,
  legend.position = "topleft"
)
}
\arguments{
\item{x}{PanelBalance object}

\item{...}{additional parameters to be passed to \code{base::plot()}}

\item{include.unrefined.panel}{logical indicating whether or not unrefined balance plots should be returned for panel plot. NEEDS TO BE IMPLEMENTED.}
}
\description{
plot covariate balance results
}
