% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelEstimate2.R
\name{plot.PanelEstimate}
\alias{plot.PanelEstimate}
\title{plot the point estimates and standard errors from a PanelEstimate calculation. The only mandatory argument is an object of the PanelEstimate class
Use standard arguments to the \code{plot} function to modify the plot as needed.}
\usage{
\method{plot}{PanelEstimate}(pe.object,
  ylab = "Estimated Effect of Treatment", xlab = "Time",
  main = "Estimated Effects of Treatment Over Time", ...)
}
\description{
plot the point estimates and standard errors from a PanelEstimate calculation. The only mandatory argument is an object of the PanelEstimate class
Use standard arguments to the \code{plot} function to modify the plot as needed.
}
