% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelBalance.R
\name{print.PanelBalance}
\alias{print.PanelBalance}
\title{Print basic information about PanelBalance objects}
\usage{
\method{print}{PanelBalance}(x, ...)
}
\arguments{
\item{x}{PanelBalance object}

\item{...}{Not used}
}
\description{
Print basic information about PanelBalance objects
}
