% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelData.R
\name{print.PanelData}
\alias{print.PanelData}
\title{Print PanelData objects and basic metadata}
\usage{
\method{print}{PanelData}(x, ...)
}
\arguments{
\item{x}{\code{PanelData} object}

\item{...}{additional arguments to be passed to \code{print.data.frame()}}
}
\description{
Print PanelData objects and basic metadata
}
\examples{
d <- PanelData(dem, "wbcode2", "year", "dem", "y")
print(d)

}
