% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelEstimateObject.R
\name{print.PanelEstimate}
\alias{print.PanelEstimate}
\title{Print PanelEstimate objects with information about estimates}
\usage{
\method{print}{PanelEstimate}(x, ..., verbose = FALSE)
}
\arguments{
\item{x}{PanelEstimate object}

\item{...}{additional arguments to be passed to \code{print.data.frame()}}

\item{verbose}{logical indicating whether or not more information about the results should be printed}
}
\description{
Print PanelEstimate objects with information about estimates
}
