% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_treated.R
\name{return_treated}
\alias{return_treated}
\title{return_treated}
\usage{
return_treated(matched_sets1, matched_sets2, lag, unit.id, qoi = "att",
  time.id, unit.name, data)
}
\arguments{
\item{matched_sets1}{a list returned by \code{PanelMatch}, 
with argument \code{naive} set to TRUE and \code{lag} set to 1}

\item{matched_sets2}{a list returned by \code{PanelMatch}}

\item{lag}{number of lags specified by the user when using \code{PanelMatch}}

\item{unit.id}{A character string indicating the name of unit identifier
variable in the \code{data}. The variable must be numeric.}

\item{qoi}{the quantity of interest}

\item{time.id}{the unit.id specified by the user when using \code{PanelMatch}}

\item{unit.name}{the name of the unit}

\item{data}{The data frame in data.frame class}
}
\value{
\code{return_treated} returns a matrix in which the first row contains the unit names
and second row contains the time.id of the treated observations that do not have matched
control units
}
\description{
\code{return_treated} obtains unit-time information for treated observations that 
do not have matched control units given 
a user-specified quantity of interest and number of lags. The user will 
first run \code{PanelMatch} with \code{naive} set to TRUE and \code{lag} 
set to 1. This will determine the total number of treated observations 
for a specific quantity of interest. The user will pass the returned object to 
argument \code{matched_sets1} in \code{return_treated}.
Then, the user will run \code{PanelMatch}
with preferred specifications, and pass the returned object to \code{matched_sets2} in \code{return_treated}
 \code{return_treated} will return the 
missing treated observations by comparing the \code{matched_sets1} and \code{matched_sets2}.
}
\author{
In Song Kim <insong@mit.edu>, Erik Wang
<haixiao@Princeton.edu>, and Kosuke Imai <kimai@Princeton.edu>
}
