% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listwise_delete.R
\name{set_lwd_refinement}
\alias{set_lwd_refinement}
\title{set_lwd_refinement
Performs the set-level operations for refinement with listwise deletion. See documentation for lwd_refinement for descriptions of most parameters.}
\usage{
set_lwd_refinement(
  mset,
  local.data,
  time,
  id,
  lag,
  refinement.method,
  lead,
  verbose,
  size.match,
  unit.id,
  time.id,
  covs.formula,
  match.missing,
  treatment,
  use.diag.covmat
)
}
\arguments{
\item{mset}{individual matched set}

\item{local.data}{data.frame containing the data relevant for set level refinement}

\item{time}{time of treated observation}

\item{id}{id of treated observation}
}
\value{
an individual matched set
}
\description{
set_lwd_refinement
Performs the set-level operations for refinement with listwise deletion. See documentation for lwd_refinement for descriptions of most parameters.
}
\keyword{internal}
