% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matched_set_obj.R
\name{[.matched.set}
\alias{[.matched.set}
\title{Subset matched.set object}
\usage{
\method{[}{matched.set}(x, i, j = NULL, drop = NULL)
}
\arguments{
\item{x}{\code{matched.set} object}

\item{i}{numeric. specifies the index of which element to extract.}

\item{j}{NULL}

\item{drop}{NULL}
}
\description{
Subsets \code{matched.set} objects while preserving attributes.
}
