% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelBalance.R
\name{[.PanelBalance}
\alias{[.PanelBalance}
\title{Subset PanelBalance objects}
\usage{
\method{[}{PanelBalance}(x, i, ...)
}
\arguments{
\item{x}{\code{PanelBalance} object}

\item{i}{numeric. Specifies which element to extract. Substantively, it specifies which \code{PanelMatch} configuration data to extract.}

\item{...}{Not used}
}
\value{
Returns balance information for specified \code{PanelMatch} configuration. Note that results are still returned as a \code{PanelBalance} object.
}
\description{
Subset PanelBalance objects
}
