% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matched_set_obj.R
\name{summary.matched.set}
\alias{summary.matched.set}
\title{summary.matched.set}
\usage{
summary.matched.set(object, ..., verbose = T)
}
\arguments{
\item{object}{a \code{matched.set} object}

\item{...}{additional arguments to be passed to \code{summary}}

\item{verbose}{logical indicating whether or not output should be in expanded form}
}
\description{
Method for summarizing \code{matched.set} objects.
}
