% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelEstimate.R
\name{summary.PanelEstimate}
\alias{summary.PanelEstimate}
\title{Title}
\usage{
summary.PanelEstimate(object)
}
\arguments{
\item{object}{takes an PanelEstimate object}
}
\value{
a summary of PanelEstimate outputs
}
\description{
Title
}
