% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{treatmentChangeDistribution}
\alias{treatmentChangeDistribution}
\title{treatmentChangeDistribution}
\usage{
treatmentChangeDistribution(pm.object, data.in, type, ...)
}
\description{
Calculates results for a placebo test
}
\details{
Plots the levels of treatment at time = t-1 and time = t
}
