% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matched_set_obj.R
\name{weights.matched.set}
\alias{weights.matched.set}
\title{Extract the weights of matched.set objects}
\usage{
\method{weights}{matched.set}(object)
}
\arguments{
\item{object}{matched.set object, extracted using the \code{get.PanelMatch()} method}
}
\value{
list of named vectors. Each list element corresponds to a particular treated observation and contains the matched control units, along with their weights.
}
\description{
Extract the weights of matched.set objects
}
