% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matched_set_obj.R
\name{weights}
\alias{weights}
\title{Get weights of matched control units
See weights.matched.set method}
\usage{
weights(object)
}
\arguments{
\item{object}{\code{Matched.set object}}
}
\description{
Get weights of matched control units
See weights.matched.set method
}
