% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dem}
\alias{dem}
\title{County-level Democracy indicator}
\format{A dataframe containing 9384 rows and 3 variables}
\source{
Acemoglu, Daron, Suresh Naidu, Pascual Restrepo, and James
A Robinson. ``Democracy does cause growth.'' Journal of Political
Economy.
}
\description{
A dataset containing the democracy indicator for 184 countries from
1960 to 2010
}
\details{
\itemize{
  \item wbcode2. World Bank country ID
  \item year. year (1960--2010)
  \item dem. binary indicator of democracy as defined in Acemoglu et al.
  \item y log of GDP per capita in 2000 constant dollars (multiplied by 100)
  \item tradewb Exports plus Imports as a share of GDP from World Bank
}
}
\keyword{dataset}
