% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisplayTreatment.R
\name{DisplayTreatment}
\alias{DisplayTreatment}
\title{DisplayTreatment}
\usage{
DisplayTreatment(unit.id, time.id, treatment, data, color.of.treated = "red",
  color.of.untreated = "blue",
  title = "Treatment Distribution \\n Across Units and Time", xlab = "Time",
  ylab = "Unit", x.size = 10, y.size = 5, legend.position = "none",
  x.angle = 45, y.angle = NULL, legend.labels = c("not treated",
  "treated"), group_on = NULL, sort_by = NULL, decreasing = FALSE,
  matched.set = NULL, show.set.only = FALSE, hide.x.axis.label = FALSE,
  hide.y.axis.label = FALSE)
}
\arguments{
\item{unit.id}{A numeric vector of unit identifiers}

\item{time.id}{A numeric vector of time identifiers}

\item{treatment}{Name of the treatment variable in character class}

\item{data}{The data frame in data.frame class}

\item{color.of.treated}{Color of the treated observations in
character. Default is red, but shade will depend on whether or not \code{group_on} is provided}

\item{color.of.untreated}{Color of the untreated observations in
character. Default is blue, but shade will depend on whether or not \code{group_on} is provided}

\item{title}{Title of the plot in character
set as in ggplot2}

\item{xlab}{Character label of the x-axis}

\item{ylab}{Character label of the y-axis}

\item{x.size}{Numeric size of the text for xlab. Default is 10. Assign x.size = NULL to use built in ggplot2 method of determining label size. 
When the length of the time period is long, consider setting to NULL and adjusting size and ratio of the plot.}

\item{y.size}{Numeric size of the text for ylab. Default is 5. Assign y.size = NULL to use built in ggplot2 method of determining label size. 
When the number of units is large, consider setting to NULL and adjusting size and ratio of the plot.}

\item{legend.position}{Position of the legend with the same choice}

\item{x.angle}{Angle (in degrees) of the tick labels for x-axis}

\item{y.angle}{Angle (in degrees) of the tick labels for y-axis}

\item{legend.labels}{Character vector of length two describing the
labels of the legend to be shown in the plot}

\item{group_on}{Character name of column with a categorical variable that is not time dependent. 
If provided, units with shared values will be grouped together and each group will be highlighted on the resulting plot.}

\item{sort_by}{Character name of a column containing data that will be used to help determine the order in which units are displayed on the plot. 
The average value for this variable will be calculated per unit. Units will then be displayed in ascending or descending order, according to that calculated value. 
Default value is the amount of treatment that units receive.}

\item{decreasing}{Logical. Determines if display order-- according to either the column specified in the \code{sort_by} argument or amount of treatment that units receive-- should be increasing or decreasing}
}
\value{
\code{DisplayTreatment} returns a treatment variation plot,
which visualizes the variation of treatment across unit and time.
}
\description{
\code{DisplayTreatment} visualizes the treatment distribution
across unit and time in a panel dataset
}
\examples{
\dontrun{
DisplayTreatment(unit.id = "wbcode2",
                 time.id = "year", legend.position = "none",
                 xlab = "year", ylab = "Country Code",
                 treatment = "dem", data = dem)
}

}
\author{
In Song Kim <insong@mit.edu>, Erik Wang
<haixiao@Princeton.edu>, and Kosuke Imai <kimai@Princeton.edu>
}
