% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_set_treatment_effects}
\alias{get_set_treatment_effects}
\title{get_set_treatment_effects}
\usage{
get_set_treatment_effects(pm.obj, data.in, lead)
}
\arguments{
\item{pm.obj}{an object of class \code{PanelMatch}}

\item{data.in}{data.frame with the original data}

\item{lead}{integer indicating the time period in the future for which the treatment effect size will be calculated. Calculations will be made for the period t + lead, where t is the time of treatment.}
}
\description{
Calculates the treatment effect size at the matched set level
}
\details{
Calculate the size of treatment effects for each matched set.
}
