% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelBCheck.R
\name{PanelBCheck}
\alias{PanelBCheck}
\title{PanelBCheck}
\usage{
PanelBCheck(matched_sets, covariate = NULL, qoi = NULL,
  post.treatment = TRUE, vline = TRUE, xlab = "Time periods",
  ylab = "Differences between the treated units \\n                        and their synthetic control units",
  legend.position = "none", theme_bw = FALSE, adjustment = TRUE,
  int_size = 0.7, plot = TRUE, linetype = "dashed", colour = "blue")
}
\arguments{
\item{matched_sets}{}

\item{covariate}{}

\item{qoi}{}

\item{post.treatment}{}

\item{vline}{}

\item{xlab}{}

\item{ylab}{}

\item{legend.position}{}

\item{theme_bw}{}

\item{adjustment}{}

\item{int_size}{}

\item{plot}{}

\item{linetype}{}

\item{colour}{}
}
\value{

}
\description{
PanelBCheck
}
