% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelBCheck.R
\name{PanelBCheck}
\alias{PanelBCheck}
\title{PanelBCheck}
\usage{
PanelBCheck(matched_sets, covariate = NULL, qoi = c("att", "atc"),
  post.treatment = TRUE, vline = TRUE, xlab = "Time periods",
  ylab = "Standardized Differences", legend.position = "none",
  theme_bw = FALSE, refinement = TRUE, plot = FALSE,
  linetype = "dashed", colour = "blue")
}
\arguments{
\item{matched_sets}{A matched set created by \code{PanelMatch}}

\item{covariate}{A character indicating the name of the covariate that the user wants to check balance of. Default is NULL, 
meaning that the function will check balance for the outcome variable.}

\item{qoi}{A character indicating the quantitaty of interest for which balance check is performed. It must be either "att" 
or "atc"}

\item{post.treatment}{A logical value indicating whether the user wants to see balance during and after the period
of treatment administration}

\item{vline}{A logical value indicating whether a vertical dashed line will be inserted in the plot to separate 
between pre-treatment and post-treatment periods.}

\item{xlab}{A character for x-axis label. The default is "Time periods"}

\item{ylab}{A character for y-axis label. The default is "Standardized Differences"}

\item{legend.position}{A character for the position of the legend on the plot}

\item{theme_bw}{Whether to have white background colour for the plot. The default is FALSE.}

\item{refinement}{A logical value indicating whether to check balance before or after the refinement. The default is TRUE, 
meaning that balance after the refinement will be shown}

\item{plot}{A logical value indicating whether the function returns a plot or returns values. 
The default is FALSE.}

\item{linetype}{A character for linetype}

\item{colour}{Colour for the vline}
}
\value{
\code{PanelBCheck} returns a list of class "list" consisting of the following components
\item{All_Balance}{The standardized difference
 between treated and control units for each matched set by time period}
\item{Interquartile_Balance}{The lowest, 1st quartile, median, 3rd quartile and the highest of these standardized
differences across matched sets by time period}
\item{Mean_Balance}{The mean standardized differences across all matched sets for each time period}
}
\description{
\code{PanelBCheck} checks balance after finding the matched sets. By default it checks balance
after refinement via methods such as Mahalanobis distance matching and propensity score matching
and weighting. The user can set \code{refinement} to FALSE, thus checking balance before refinement.
Balance is assessed by taking the difference between the values of the 
user-specified covairate (or outcocme) in the treated unit and the weighted
average of that across all the control units in each matched set, divided by 
one standard deviation of the values of the user-specified covariate (or outcome) across all 
treated units of all matched sets. If \code{refinement} is FALSE, then each control unit 
will receive equal weight. If \code{refinement} is TRUE, then each control unit will
receive a weight produced by the refinement process.
}
\examples{
\dontrun{
matches.cbps <- PanelMatch(lag = 4, max.lead = 4, time.id = "year",
unit.id = "wbcode2", treatment = "dem",
formula =  y ~ dem, method = "Pscore",
weighting = TRUE,  qoi = "ate",  M = 5, data = dem)
PanelBCheck(matches.cbps, qoi = "att")
}
}
\author{
In Song Kim <insong@mit.edu>, Erik Wang
<haixiao@Princeton.edu>, and Kosuke Imai <kimai@Princeton.edu>
}
