% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelBCheck.R
\name{PanelBCheck}
\alias{PanelBCheck}
\title{PanelBCheck}
\usage{
PanelBCheck(matched_sets, covariate = NULL, qoi = NULL,
  post.treatment = TRUE, vline = TRUE, xlab = "Time periods",
  ylab = "Standardized Mean Differences", legend.position = "none",
  theme_bw = FALSE, adjustment = TRUE, int_size = 0.7, plot = TRUE,
  linetype = "dashed", colour = "blue")
}
\arguments{
\item{matched_sets}{d}

\item{covariate}{d}

\item{qoi}{d}

\item{post.treatment}{d}

\item{vline}{d}

\item{xlab}{d}

\item{ylab}{d}

\item{legend.position}{d}

\item{theme_bw}{d}

\item{adjustment}{d}

\item{int_size}{d}

\item{plot}{d}

\item{linetype}{d}

\item{colour}{d}
}
\value{
\code{PanelBCheck} returns a balance results
}
\description{
PanelBCheck
}
\examples{
\dontrun{
a <- 3
}
}
