% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelEstimate.R
\name{PanelEstimate}
\alias{PanelEstimate}
\title{PanelEstimate}
\usage{
PanelEstimate(lead, inference = c("wfe", "bootstrap"), ITER = 1000,
  estimator = "did", df.adjustment = FALSE, qoi = NULL, CI = 0.95, sets,
  data, outcome.variable)
}
\arguments{
\item{lead}{An integer vector indicating the sequence of the lead
periods for which the quantity of interest will be estimated. When \code{inference} is "wfe", you can only specify one lead period at a time.}

\item{inference}{One of ``wfe'' (weighted fixed effects) or
``bootstrap'' methods for standard error calculation. The default
is \code{bootstrap}.}

\item{ITER}{An integer value indicating the number of bootstrap
iteration. The default is 1000.}

\item{estimator}{One of \code{did} (difference-in-differences) or
\code{matching} specifying the estimator for WFE regressions. The default is
\code{did}.}

\item{df.adjustment}{A logical value indicating whether a
degree-of-freedom adjustment should be performed for standard error
calculation. The default is \code{FALSE}.}

\item{qoi}{One of ``att'' (average treatment effect for the
treated) or ``ate'' (average treatment effect) or atc (average
treatment effect for the control).}

\item{CI}{A numerical value specifying the range of interval
estimates for statistical inference. The default is .95.}

\item{sets}{A list of class `matched.set' attained by
\code{PanelMatch}. @seealso \code{PanelMatch}.}

\item{data}{The same time series cross sectional data set provided to the PanelMatch function to produce the \code{sets}}
}
\value{
\code{PanelEstimate} returns a list of class
`PanelEstimate' containing the following components:
\item{coefficients}{the point estimates of the quantity of interest}
\item{bootstrapped.coefficients}{the bootstrapped coefficients, if applicable}
\item{bootstrap.iterations}{the number of iterations, if applicable}
\item{method}{refinement method used to create the matched sets from which the estimates were calculated}
\item{lag}{See PanelMatch argument \code{lag} for more information.}
\item{lead}{The lead window sequence for which PanelEstimate is producing point estimates and standard errors.}
\item{confidence.level}{the confidence interval range}
\item{qoi}{the quantity of interest}
\item{matched.sets} the refined matched sets used to produce the estimations. These might be different from the provided 
\code{sets} argument depending on the missingness of data in the lead window and the \code{qoi}
\item{df}{if \code{inference} is "wfe", the degrees of freedom}
\item{standard.error} {the standard error of the point estimates}
}
\description{
\code{PanelEstimate} estimates causal quantity of interests, e.g.,
the average treatment effect of for the treated (ATT), by
estimating the counterfactual outcomes for each treated unit using
a matched set. Users will specify matched sets that are obtained by the
\code{PanelMatch} function and obtain point estimates via weighted fixed effects regressions or via
weighted average computation with weighted bootstrap standard errors
}
\examples{
\dontrun{
pm.obj <- PanelMatch(4, "year", "wbcode2", "dem", "y", refinement.method = "mahalanobis", data = dem, match.missing = T, covs.formula = ~ tradewb + lag("tradewb", 1:4) + lag("y", 1:4), size.match = 5)
res <- PanelEstimate(lead = 0:4, inference = "bootstrap", qoi = "att", sets = pm.obj, data = dem, outcome.variable = "y")
res2 <- PanelEstimate(lead = 0, inference = "wfe", qoi = "att", sets = pm.obj, data = dem, outcome.variable = "y")
}
}
\author{
In Song Kim <insong@mit.edu>, Erik Wang
<haixiao@Princeton.edu>, and Kosuke Imai <kimai@Princeton.edu>
}
