% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelEstimate.R
\name{PanelEstimate}
\alias{PanelEstimate}
\title{PanelEstimate}
\usage{
PanelEstimate(inference = c("wfe", "bootstrap"), ITER = 1000,
  df.adjustment = FALSE, CI = 0.95, sets, data)
}
\arguments{
\item{inference}{One of ``wfe'' (weighted fixed effects) or
``bootstrap'' methods for standard error calculation. The default
is \code{bootstrap}.}

\item{ITER}{An integer value indicating the number of bootstrap
iteration. The default is 1000.}

\item{df.adjustment}{A logical value indicating whether a
degree-of-freedom adjustment should be performed for standard error
calculation. The default is \code{FALSE}.}

\item{CI}{A numerical value specifying the range of interval
estimates for statistical inference. The default is .95.}

\item{sets}{A list of class `PanelMatch' attained by
\code{PanelMatch}.}

\item{data}{The same time series cross sectional data set provided to the PanelMatch function to produce the \code{sets}}
}
\value{
\code{PanelEstimate} returns a list of class
`PanelEstimate' containing the following components:
\item{coefficients}{the point estimates of the quantity of interest}
\item{bootstrapped.coefficients}{the bootstrapped coefficients, if applicable}
\item{bootstrap.iterations}{the number of iterations, if applicable}
\item{method}{refinement method used to create the matched sets from which the estimates were calculated}
\item{lag}{See PanelMatch argument \code{lag} for more information.}
\item{lead}{The lead window sequence for which PanelEstimate is producing point estimates and standard errors.}
\item{confidence.level}{the confidence interval range}
\item{qoi}{the quantity of interest}
\item{matched.sets}{the refined matched sets used to produce the estimations}
\item{df}{if \code{inference} is "wfe", the degrees of freedom}
\item{standard.error}{the standard error of the point estimates}
}
\description{
\code{PanelEstimate} estimates a causal quantity of interest, including the average treatment effect for treated or control units (att and atc, respectively), or average treatment effect (ate), as specified in \code{PanelMatch}
This is done by estimating the counterfactual outcomes for each treated unit using
matched sets. Users will specify matched sets that were obtained by the
\code{PanelMatch} function and obtain point estimates via weighted fixed effects regressions or via
weighted average computation with weighted bootstrap standard errors.
}
\examples{
\dontrun{
PM.results <- PanelMatch(lag = 4, time.id = "year", unit.id = "wbcode2", 
                         treatment = "dem", refinement.method = "mahalanobis", 
                         data = dem, match.missing = T, 
                         covs.formula = ~ lag("tradewb", 1:4) + lag("y", 1:4), 
                         size.match = 5, qoi = "att",
                         outcome.var = "y", lead = 0:4, forbid.treatment.reversal = TRUE)
PE.results <- PanelEstimate(inference = "bootstrap", sets = PM.results, data = dem)

PM.results <- PanelMatch(lag = 4, time.id = "year", unit.id = "wbcode2", 
                         treatment = "dem", refinement.method = "mahalanobis", 
                         data = dem, match.missing = T, 
                         covs.formula = ~ lag("tradewb", 1:4) + lag("y", 1:4), 
                         size.match = 5, qoi = "att",
                         outcome.var = "y", lead = 0, forbid.treatment.reversal = TRUE)
PE.results <- PanelEstimate(inference = "wfe", sets = PM.results, data = dem)
}
}
\author{
In Song Kim <insong@mit.edu>, Erik Wang
<haixiao@Princeton.edu>, Adam Rauh <adamrauh@mit.edu>, and Kosuke Imai <kimai@Princeton.edu>
}
