% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelMatch-package.R
\docType{package}
\name{PanelMatch-package}
\alias{PanelMatch}
\alias{PanelMatch-package}
\title{Matching Methods for Causal Inference with Time-Series Cross-Section Data}
\description{
Matching Methods for Causal Inference with Time-Series Cross-Section Data
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/insongkim/PanelMatch/issues}
}

}
\author{
\strong{Maintainer}: In Song Kim \email{insong@mit.edu}

Authors:
\itemize{
  \item Erik Wang \email{haixiaow@Princeton.edu}
  \item Kosuke Imai \email{kimai@Princeton.edu}
}

}
