% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelMatch.R
\name{PanelMatch}
\alias{PanelMatch}
\title{Illustration of PanelMatch}
\usage{
PanelMatch(lag, max.lead, time.id = "year", qoi = "ate",
  unit.id = "ccode", treatment, formula = y ~ treat, data,
  weighting = FALSE, M = 3, covariate.only = FALSE, method = NULL)
}
\description{
\code{PanelMatch} finds matched set for each treated observation  with the number of lags and forwards as well as covariates for matching and weighting refinement determined by the user
}
