% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelMatch.R
\name{PanelMatch}
\alias{PanelMatch}
\title{PanelMatch}
\usage{
PanelMatch(lag, max.lead, time.id = "year", qoi = "ate",
  unit.id = "ccode", treatment, formula = y ~ treat, data,
  weighting = FALSE, M = 3, covariate.only = FALSE, method = NULL)
}
\arguments{
\item{lag}{An integer value indicating the length of history to be matched}

\item{max.lead}{An optional integer value indicating the length of maximum leads}

\item{time.id}{A character string indicating the name of time
variable in the \code{data}}

\item{qoi}{One of ``att'' (average treatment effect for the
treated) or ``ate'' (average treatment effect) or atc (average
treatment effect for the control). The default is ``att.''}

\item{unit.id}{A character string indicating the name of unit
variable in the \code{data}}

\item{treatment}{A character string indicating the name of
treatment variable in the \code{data}. The treatment should be a
binary indicator (integer with 0 for the control group and 1 for
the treatment group).}

\item{formula}{A symbolic description of the outcome variable and
covariates whose history should be used to refine the matched set.}

\item{data}{The data frame in data.frame class}

\item{weighting}{A logical value indicating whether refinement
through weighting should be performed. The default is \code{FALSE}.}

\item{M}{An integer value indicating the maximum number of the
control units in each matched set. The default is 3.}

\item{covariate.only}{A logical value indicating whether the
history should be matched only based on covariates. The default is
\code{FALSE}.}

\item{method}{An optional character string indicating the matching
methods for refining the matched set based on covariates. One of
``Maha'', ``Pscore'', ``CBPS'', ``SynthPscore'', and ``SynthCBPS''.}
}
\value{
\code{PanelMatch} returns a list of class `panelmatch'
containing the following components:
\item{treatment}{treatment variable name}
\item{qoi}{the quantity of interest}
\item{dependent}{the dependent variable}
\item{covariate}{the covariates}
\item{unit.id}{the unit id}
\item{time.id}{the time id}
\item{M}{the length of history to be matched}
\item{covariate.only}{the indicator of whether only covariates are matched}
\item{lag}{the length of lags}
\item{max.lead}{the length of maximum lead}
\item{data}{the data frame}
\item{ATT_matches}{the matched set given the att as the quantity of interest}
\item{ATC_matches}{the matched set given the atc as the quantity of interest}
\item{covariate_names}{Names of covariates in the user-specified formula}
\item{NC_ATT}{...}
\item{NC_ATC}{...}
}
\description{
\code{PanelMatch} identifies a matched set for each treated
observation. Specifically, for a given treated unit, the matched
set consists of control observations that have the identical
treatment history up to a certain number of \code{lag}
years. Researchers must specify \code{lag}. A further refinement of
the matched set will be possible by setting the size of the matched
set \code{M}, \code{weighting}, and adjusting for other confounders
such as past outcomes and covariates via \code{formula}.
}
\examples{
\dontrun{
matches.cbps <- PanelMatch(lag = 4, max.lead = 4,
                           time.id = "year", unit.id = "wbcode2",
                           treatment = "dem", formula = y ~ dem,
                           method = "CBPS", weighting = FALSE,
                           qoi = "ate", M = 5, data = dem)
}
}
\author{
In Song Kim <insong@mit.edu>, Erik Wang
<haixiao@Princeton.edu>, and Kosuke Imai <kimai@Princeton.edu>
}
