% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelMatch.R
\name{PanelMatch}
\alias{PanelMatch}
\title{PanelMatch}
\usage{
PanelMatch(lag, time.id, unit.id, treatment, refinement.method,
  size.match = 10, data, match.missing = TRUE, covs.formula = NULL,
  verbose = FALSE, qoi, lead = 0, outcome.var,
  exact.match.variables = NULL, forbid.treatment.reversal = FALSE,
  matching = TRUE, listwise.delete = FALSE,
  use.diagonal.variance.matrix = FALSE)
}
\arguments{
\item{lag}{An integer value indicating the length of history to be matched}

\item{time.id}{A character string indicating the name of time identifier
variable in the \code{data}. This data currently must be integers that increase by one.}

\item{unit.id}{A character string indicating the name of unit identifier in the data. This data must be character, integer, or numeric. However, it is recommended to use integers as ids.}

\item{treatment}{A character string indicating the name of treatment variable in the \code{data}. The treatment must be a binary indicator (integer with 0 for the control group and 1 for the treatment group).}

\item{refinement.method}{character string of matching or weighting method used for refining the matched sets. The user can choose "mahalanobis", "ps.match", "CBPS.match", "ps.weight", "CBPS.weight", "ps.msm.weight", "CBPS.msm.weight", or "none". The first three methods will use the \code{size.match} argument to create sets of at most \code{size.match} closest control units. Choosing "none" will assign equal weights to all control units in each matched sets.}

\item{size.match}{Maximum size of the matched sets after refinement. This argument only affects results when using a matching method (any of the refinement methods that end in .match). This argument is not needed and will have no impact if included on a weighting method.}

\item{data}{A data.frame object containing time series cross sectional data. Time data must be integers that increase by 1.}

\item{match.missing}{Logical variable indicating whether or not units should be matched on the patterns of missingness in their treatment histories}

\item{covs.formula}{One sided formula indicating which variables should be used for matching and refinement. Argument is optional if \code{refinement.method} is set to "none"
If the user wants to include lagged variables, this can be done using a function, "lag()", which takes two, unnamed, positional arguments. The first is the name of the variable which you wish to lag, specified as a string. The second is the lag window, specified as an integer sequence
For instance, lag("x", 1:4) will then add new columns to the data for variable "x" for time t-1, t-2, t-3, and t-4.}

\item{verbose}{option to include more information about the matched.set object calculations, like the distances used to create the refined sets and weights.}

\item{qoi}{quantity of interest: att (average treatment effect on treated units), atc (average treatment effect on control units), ate (average treatment effect). Note that the qoi for MSM methods will give the estimated average treatment effect of being treated for a chosen F time periods. This differs slightly from the non-MSM methods -- treatment is not required for F periods for those methods.}

\item{lead}{integer sequence specifying the lead window for which qoi estimates will ultimately be produced. Default is 0.}

\item{outcome.var}{Character string of the outcome variable.}

\item{exact.match.variables}{character vector giving the names of variables to be exactly matched on. These should be time invariant variables.}

\item{forbid.treatment.reversal}{Logical indicating whether or not it is permissible for treatment to reverse. This must be set to TRUE for msm methods. When set to TRUE, only matched sets where treatment is applied continuously are included.}

\item{matching}{logical indicating whether or not any matching on treatment history should be performed. This is used for diagnostic purposes. Default is TRUE.}
}
\value{
\code{PanelMatch} returns an object of class "PanelMatch". This is a list that contains a few specific elements: First, a matched.set object(s) that has the same name as the provided qoi if the qoi is "att" or "atc". 
If qoi = "ate" then two matched.set objects will be attached, named "att" and "atc." This object also has some additional attributes:
\item{qoi}{The qoi specified in the original function call}
\item{lead}{the lead window specified in the original function call}
\item{forbid.treatment.reversal}{logial value matching the forbid.treatment.reversal parameter provided in the function call.}
\item{outcome.var}{character string matching the outcome variable provided in the original function call.}
For more information about matched.set objects, see documentation for the "matched_set" function
}
\description{
Create refined/weighted sets of treated and control units
}
\details{
\code{PanelMatch} identifies a matched set for each treated
observation. Specifically, for a given treated unit, the matched
set consists of control observations that have the identical
treatment history up to a certain number of \code{lag}
years. Researchers must specify \code{lag}. A further refinement of
the matched set will be possible by setting a maximum size of the matched
set \code{size.match} (the maximum number of control units that can be matched to a treated unit) and adjusting for other confounders
such as past outcomes and covariates via \code{covs.formula} and specifying a method of refinement via \code{refinement.method}.
}
\examples{
\dontrun{
PM.results <- PanelMatch(lag = 4, time.id = "year", unit.id = "wbcode2", 
                         treatment = "dem", refinement.method = "mahalanobis", 
                         data = dem, match.missing = T, 
                         covs.formula = ~ lag("tradewb", 1:4) + lag("y", 1:4), 
                         size.match = 5, qoi = "att",
                         outcome.var = "y", lead = 0:4, forbid.treatment.reversal = FALSE)
}
}
\author{
Adam Rauh <adamrauh@mit.edu>, In Song Kim <insong@mit.edu>, Erik Wang
<haixiao@Princeton.edu>, and Kosuke Imai <kimai@Princeton.edu>
}
