% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelMatch2.R
\name{PanelMatch2}
\alias{PanelMatch2}
\title{PanelMatch2}
\usage{
PanelMatch2(lag, time.id, unit.id, treatment, outcome,
  refinement.method = c(NULL, "ps.weight", "ps.match", "mahalanobis",
  "CBPS.weight", "CBPS.match"), size.match = 10, data, match.missing = TRUE,
  covs.formula, verbose = FALSE)
}
\arguments{
\item{lag}{An integer value indicating the length of history to be matched}

\item{time.id}{A character string indicating the name of time identifier
variable in the \code{data}. currently must be integer.}

\item{unit.id}{A character string indicating the name of unit identifier in the data. Currently must be integer.}

\item{treatment}{A character string indicating the name of
treatment variable in the \code{data}. The treatment should be a
binary indicator (integer with 0 for the control group and 1 for
the treatment group).}

\item{outcome}{Character string of outcome variable.}

\item{refinement.method}{character string of matching method used for refining the matched sets. Currently only mahalanobis is supported}

\item{size.match}{Maximum size of the matched sets after refinement}

\item{data}{Balanced panel data set}

\item{match.missing}{Logical variable indicating whether or not units can be matched on the patterns of missingness in their treatment histories}

\item{covs.formula}{One sided formula indicating which variables should be used for matching and refinement.
Can specify lags using a function "lag" which takes two positional arguments. The first is the name of the variable which you wish to lag, specified as a string second is the lag window, specified as an integer sequence
See the example below.}

\item{formula}{A symbolic description of the outcome variable and
only covariates with contemporaneous effects.}
}
\value{
\code{PanelMatch} returns a list of class `matched.set'
containing the following components:
\item{treatment}{treatment variable name}
}
\description{
\code{PanelMatch2} identifies a matched set for each treated
observation. Specifically, for a given treated unit, the matched
set consists of control observations that have the identical
treatment history up to a certain number of \code{lag}
years. Researchers must specify \code{lag}. A further refinement of
the matched set will be possible by setting the size of the matched
set \code{M}, \code{weighting}, and adjusting for other confounders
such as past outcomes and covariates via \code{formula}.
}
\examples{
\dontrun{
results <- PanelMatch2(4, "year", "wbcode2", "dem", "y", refinement.method = "mahalanobis", 
                      data = dem, match.missing = T, covs.formula = ~ lag("tradewb", 1:4) + lag("y", 1:4), size.match = 5, verbose = T)
}
}
\author{
In Song Kim <insong@mit.edu>, Erik Wang
<haixiao@Princeton.edu>, and Kosuke Imai <kimai@Princeton.edu>
}
