% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matched_set_obj.R
\name{plot.matched.set}
\alias{plot.matched.set}
\title{Plot the distribution of the sizes of matched sets.}
\usage{
\method{plot}{matched.set}(x, ..., border = NA, col = "grey",
  ylab = "Frequency of Size", xlab = "Matched Set Size", lwd = NULL,
  main = "Distribution of matched set sizes", freq = TRUE,
  include.empty.sets = TRUE)
}
\arguments{
\item{x}{a \code{matched.set} object}

\item{...}{optional arguments to be passed to \code{plot}}

\item{border}{default is NA. This is the same argument as the standard argument for \code{plot}}

\item{col}{default is "grey". This is the same argument as the standard argument for \code{plot}}

\item{ylab}{default is "Frequency of Size". This is the same argument as the standard argument for \code{plot}}

\item{xlab}{default is "Matched Set Size". This is the same argument as the standard argument for \code{plot}}

\item{lwd}{default is NULL. This is the same argument as the standard argument for \code{plot}}

\item{main}{default is "Distribution of matched set sizes". This is the same argument as the standard argument for \code{plot}}

\item{freq}{default is TRUE. See \code{freq} argument in \code{hist} function for more}

\item{include.empty.sets}{default is TRUE. Should empty sets be included on the plot? If false, they will be excluded, but noted as a subtitle.}
}
\description{
A plot method for creating histograms of the distribution of the sizes of matched sets. 
Accepts all standard \code{plot} arguments. Empty sets are excluded from the histogram but are noted in the subtitle by default.
}
\examples{
\dontrun{
PM.results <- PanelMatch(lag = 4, time.id = "year", unit.id = "wbcode2", 
                         treatment = "dem", refinement.method = "mahalanobis", 
                         data = dem, match.missing = T, 
                         covs.formula = ~ I(lag(tradewb, 1:4)) + I(lag(y, 1:4)),
                         size.match = 5, qoi = "att",
                         outcome.var = "y", lead = 0:4, forbid.treatment.reversal = FALSE)
plot(PM.results$att)
}

}
