% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matched_set_obj.R
\name{print.matched.set}
\alias{print.matched.set}
\title{Method for printing \code{matched.set} objects.}
\usage{
print.matched.set(x, ..., verbose = F)
}
\arguments{
\item{x}{a \code{matched.set} object}

\item{...}{additional arguments to be passed to \code{print}}

\item{verbose}{logical indicating whether or not output should be printed in expanded form}
}
\description{
Method for printing \code{matched.set} objects.
}
