% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelBalance.R
\name{summary.PanelBalance}
\alias{summary.PanelBalance}
\title{Summarize covariate balance over time (placeholder for now)}
\usage{
\method{summary}{PanelBalance}(
  object,
  qoi = NULL,
  include.unrefined = TRUE,
  unrefined.only = FALSE,
  ...
)
}
\arguments{
\item{qoi}{Character string "att", "art", or "atc" -- specifying which qoi to extract.}

\item{...}{Not used}

\item{PanelBalance}{object}

\item{refined}{logical. return refined or unrefined balance results. Default is TRUE.}
}
\description{
Summarize covariate balance over time (placeholder for now)
}
