% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelData.R
\name{summary.PanelData}
\alias{summary.PanelData}
\title{Summarize information about variable names, and unit, time, and treatment data in a PanelData object}
\usage{
\method{summary}{PanelData}(object, ...)
}
\arguments{
\item{object}{\code{PanelData} object}

\item{...}{Not used}
}
\value{
Returns a \code{data.frame} object, with columns "quantity" and "value." Within the data frame the following information is returned: The name of the unit id variable, the name of the time id variable, the name of the treatment variable, the name of the outcome variable, the number of unique units found in the data, the number of unique time periods found in the data and the percentage of treated periods that are missing treatment data.
}
\description{
Summarize information about variable names, and unit, time, and treatment data in a PanelData object
}
\examples{
d <- PanelData(dem, "wbcode2", "year", "dem", "y")
summary(d)

}
