% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelData.R
\name{summary.PanelData}
\alias{summary.PanelData}
\title{Summarize the number of unique units and time periods in a PanelData object}
\usage{
\method{summary}{PanelData}(object, ...)
}
\arguments{
\item{object}{PanelData object}

\item{...}{Not used}
}
\value{
Returns a \code{data.frame} object, with columns "num.units" and "num.periods." These specify the number of unique units and time periods that appear in the balanced panel data.
}
\description{
Summarize the number of unique units and time periods in a PanelData object
}
\examples{
d <- PanelData(dem, "wbcode2", "year", "dem", "y")
summary(d)

}
