% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelEstimate2.R
\name{summary.PanelEstimate}
\alias{summary.PanelEstimate}
\title{Get summaries of PanelEstimate objects}
\usage{
\method{summary}{PanelEstimate}(object, ...)
}
\arguments{
\item{object}{A PanelEstimate object}

\item{...}{Further arguments to be passed to \code{summary.PanelEstimate()}.}
}
\description{
\code{summary.PanelEstimate()} takes an object returned by
\code{PanelEstimate}, and returns a summary table of point
estimates and the confidence intervales.
}
