% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelEstimate.R
\name{summary.PanelEstimate}
\alias{summary.PanelEstimate}
\title{Get summaries of PanelEstimate objects/calculations}
\usage{
\method{summary}{PanelEstimate}(object, verbose = TRUE,
  bias.corrected = FALSE, ...)
}
\arguments{
\item{object}{A PanelEstimate object}

\item{verbose}{logical indicating whether or not output should be printed in an expanded form. Default is TRUE}

\item{bias.corrected}{logical indicating whether or not bias corrected estimates should be provided. Default is FALSE}

\item{...}{optional additional arguments. Currently, no additional arguments are supported.}
}
\description{
\code{summary.PanelEstimate} takes an object returned by
\code{PanelEstimate}, and returns a summary table of point
estimates and confidence intervals
}
\examples{
PM.results <- PanelMatch(lag = 4, time.id = "year", unit.id = "wbcode2", 
                         treatment = "dem", refinement.method = "mahalanobis", 
                         data = dem, match.missing = TRUE, 
                         covs.formula = ~ I(lag(tradewb, 1:4)) + I(lag(y, 1:4)),
                         size.match = 5, qoi = "att",
                         outcome.var = "y", lead = 0:4, forbid.treatment.reversal = FALSE)
PE.results <- PanelEstimate(sets = PM.results, data = dem)
summary(PE.results)



}
